+++ src/messenger/messenger_list.ts
@@ -5,20 +5,8 @@ import { assertIsNotMainThread } from '@utilities';
 
 import type { Messenger } from './messenger.js';
 import type { BaseWorkerData } from '@typing/worker_data.js';
+import type { MessengerList as MessengerListType } from '@typing/messenger.js';
 
-/**
- * For use only within workers.
- *
- * @param name The name of the `Messenger` to use.
- * @returns A promise of a {@link Messenger} instance that can be used to send messages between threads.
- *
- * This function will throw an error if a `Messenger` instance with the specified name was never sent to the worker.
- *
- * @example
- * const messenger = await MessengerList.use('foo');
- *
- * messenger.onMessage<string>((data) => console.log(data, 'received!'));
- */
 async function use(name: string) {
     assertIsNotMainThread('MessengerList.use');
     const { messengers } = workerData as BaseWorkerData;
@@ -41,20 +29,8 @@ async function use(name: string) {
     return messenger;
 }
 
-/**
- * Grab hold of an object containing all `Messenger`s available within the current worker.
- *
- * **Tip:** Useful when debugging issues involving `Messenger`s.
- *
- * @returns An object containing `Messenger`s organized based on their names.
- *
- * @example
- * const map = MessengerList.seek();
- * console.log(map); // -> { foo: Messenger, bar: Messenger }
- * console.log(map.foo.uniqueKey);
- */
 function list() {
-    assertIsNotMainThread('MessengerList.seek');
+    assertIsNotMainThread('MessengerList.list');
     return (workerData as BaseWorkerData).messengers;
 }
 
@@ -66,7 +42,15 @@ function list() {
  *
  * messenger.onMessage<string>((data) => console.log(data, 'received!'));
  */
-export const MessengerList = {
-    list,
-    use,
-};
+export const MessengerList: MessengerListType = Object.create(
+    {
+        use,
+    },
+    {
+        list: {
+            get() {
+                return list();
+            },
+        },
+    }
+);