+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- `MessengerList.list()` to be a getter method instead of a regular function. New syntax is `MessengerList.list`.
+
 ## [0.3.0] - 2022-29-12
 
 ### Removed