+++ CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - `MessengerList.list()` to be a getter method instead of a regular function. New syntax is `MessengerList.list`.
+- `SharedMap` to use `.subarray()` instead of `.slice()` for a slight memory complexity improvement.
+
+### Added
+
+- New sections to README for properties & methods of `Service`, `ServiceCluster`, and `pool`.
 
 ## [0.3.0] - 2022-29-12